Move builtins list and args from Functions dir
