+++ lib/Runtime/Library/Chakra.Runtime.Library.vcxproj
@@ -52,6 +52,8 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="$(MSBuildThisFileDirectory)ArgumentsObject.cpp" />
+    <ClCompile Include="$(MSBuildThisFileDirectory)ArgumentsObjectEnumerator.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ArrayBuffer.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Array\ES5Array.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Array\ES5ArrayIndexEnumerator.cpp" />
@@ -72,10 +74,7 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)EngineInterfaceObject.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ExternalLibraryBase.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)ForInObjectEnumerator.cpp" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)Functions\ArgumentsObject.cpp" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)Functions\ArgumentsObjectEnumerator.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Functions\BoundFunction.cpp" />
-    <ClCompile Include="$(MSBuildThisFileDirectory)Functions\JavascriptBuiltInFunctions.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Functions\JavascriptExternalFunction.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Functions\JavascriptFunction.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)Functions\RuntimeFunction.cpp" />
@@ -106,6 +105,7 @@
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptBigIntObject.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptBoolean.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptBooleanObject.cpp" />
+    <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptBuiltInFunctions.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptDate.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptError.cpp" />
     <ClCompile Include="$(MSBuildThisFileDirectory)JavascriptExceptionMetadata.cpp" />
