+++ lib/Runtime/Library/JavascriptBuiltInFunctions.cpp
@@ -11,14 +11,14 @@ namespace Js
     // Declare all the entry infos
 #define BUILTIN(c, n, e, i) FunctionInfo c::EntryInfo::n(FORCE_NO_WRITE_BARRIER_TAG(c::e), (Js::FunctionInfo::Attributes)(i), JavascriptBuiltInFunction:: ## c ## _ ## n);
 #define BUILTIN_TEMPLATE(c, n, e, i) template<> BUILTIN(c, n, e, i)
-#include "Library/Functions/JavascriptBuiltInFunctionList.h"
+#include "JavascriptBuiltInFunctionList.h"
 #undef BUILTIN
 
 
     FunctionInfo * const JavascriptBuiltInFunction::builtInFunctionInfo[MaxBuiltInEnum] =
     {
     #define BUILTIN(c, n, e, i) &c::EntryInfo::n,
-    #include "Library/Functions/JavascriptBuiltInFunctionList.h"
+    #include "JavascriptBuiltInFunctionList.h"
     #undef BUILTIN
     };
